requirements.txt
hydra-core
omegaconf
torch
numpy
tqdm
matplotlib
einops
scipy